Fire
I was expecting him to pull it all together in one big file and kept referring to them as patches, so I made patches for each. And then he just did play sound for each one, which you can't do with the patches. So I just made more files for it. Annoying but I'm not too worried about cleaning this project up.